refactor(integration): merge integration handlers and dto
